requirements.txt
aiogram==3.3.0
tortoise-orm==0.20.0
tortoise-orm[aiosqlite]
tortoise-orm[asyncpg]
redis==5.0.1
python-dotenv==1.0.0
dataclasses-json==0.6.3
openpyxl==3.1.2
aiogoogle==5.5.0
aiogram_dialog==2.2.0a3
asyncpg==0.29.0